Optimize Performance
(1) getTransformProperties was unnecessarily causing page reflows.
Instead of relying on this function to get computed transform
properties, transform properties are now simply cached in the class
properties #bounceCurrentTranslateX and #bounceCurrentTranslateY as
bouncing occurs.

(2) Added new property #allowCursorSwitching and a setter for it,
setAllowCursorSwitching. This determines whether pointerdown and
pointerup events cause the cursor to change, i.e. grab -> grabbing, etc.
This new property is set to true by default, but it can be set to false
if cursor switching is not wanted. Unfortunately, switching cursors
triggers page reflows, which can lead to noticeable performance issues
on large pages.