Update MomentumScroller Events and Handlers
(1) The "momentumScrollerRoute" event had 2 handlers, but they are now
consolidated. A new condition was added: when a MomentumScroller
instance's scrollContainer does not match the routeTarget and it is not
currently handling the pointer and it has a scrollResolve, the instance
is forced to resolve, and thus stop, the ongoing scroll.

(2) The MomentumScroller private class field #pointerIsDown was renamed
to #isCurrentlyHandlingPointer for clarity.

(3) The ScrollContainerTools.getScrollableAxes method parameter
ignoreElementsWithOverflowHidden was renamed
considerOverflowHiddenAxesNonScrollable for clarity.

(4) The instance setters setAllowHorizontalScrolling and
setAllowVerticalScrolling now change the elements overflow property.
When set to true, the respective axis's overflow is set to "auto"; when
set to false, the respective axis's overflow is set to hidden.

(5) Brought back "momentumScrollerPointerDown" and
"momentumScrollerPointerUp" events but renamed them to
"momentumScrollerPointerHandlingStart" and
"momentumScrollerPointerHandlingStop", respectively. These were brought
back because regular "pointerdown" and "pointerup"/"pointercancel" were
not reliably firing. "gotpointercapture" and "lostpointercapture" also
did not reliably fire due to the fact that the spec requires that they
not fire immediately after setting/releasing pointer capture, but rather
after the next pointer event. For more details about this, see:
https://github.com/w3c/pointerevents/issues/133