Update Event Handlers, Router, etc.
(1) autoCreateScrollers now creates a scroller if an element is
potentially scrollable, not just currently scrollable.

(2) undoPointerDownChanges has a new parameter, "interruptedBy". Its
argument is passed to the momentaMouseScrollerPointerHandlingStop
event dispatch as part of the CustomEvent detail object.

(3) "contextmenu" was added to an event listener within the
thresholdTest for abort purposes.

(4) event.stopPropagation() and capture mode were added to the "click"
event listener for Gecko.