Allow MomentaMouse To Work On Touch-Screen Devices
(1) MomentaMouse now works as long as a mouse is detected. It even works
on touch-screens if a mouse is available. The page now uses CSS media
queries as heuristics to identify if a fine, hoverable pointer is
available. Although this is not mouse-specific (it may also accept
touchpads), it is the best solution for now. When a mouse is found,
MomentaMouse automatically activates. When a mouse is not found,
MomentaMouse disables itself.

(2) Removed #inactiveScrollers and #activeScrollers Sets because so far
the logic has not benefited from them versus just using #scrollerMap

(3) Added activationReason and deactivationReason to the activate and
deactivate instance methods, respectively. These reasons are included
in their respective CustomEvents that are dispatched so that listeners
may react appropriately.

(4) The _pointerDownRouter static method now filters out PointerEvents
that do not have a pointerType of "mouse"

(5) The wheel listeners were made passive

(6) Removed "cancelable: true" from CustomEvents