Add Support for Coalesced Events
(1) pointermove events may be coalesced, i.e. throttled, but browsers,
meaning that some events are dropped. For applications that require
a lot of precision, such as MomentaMouse, coalesced events are not
ideal. Therefore, MomentaMouse now checks if the browser supports
PointerMove.getCoalescedEvents, and if it does, it uses all available
events that would have otherwise been coalesced.